requirements.txt
asgiref==3.8.1
Django==6.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
python-dotenv==1.0.1
sqlparse==0.5.0
whitenoise==6.6.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
